requirements.txt
amqp==5.3.1
async-timeout==5.0.1
bcrypt==4.0.1
beautifulsoup4==4.13.5
billiard==4.2.1
blinker==1.9.0
# celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cssutils==2.11.1
cycler==0.12.1
et_xmlfile==2.0.0
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-WTF==1.1.1
fonttools==4.59.2
greenlet==3.2.4
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
kombu==5.5.4
xhtml2pdf
MarkupSafe==3.0.2
matplotlib==3.9.4
more-itertools==10.8.0
mysql-connector-python==8.1.0
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prompt_toolkit==3.0.52
protobuf==4.21.12
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2025.2
# redis==6.4.0
seaborn==0.13.2
sendgrid==6.10.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
starkbank-ecdsa==2.2.0
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.3.7
WTForms==3.2.1
zipp==3.23.0
gunicorn
